CRM-DocumentIA/rag_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pinecone
sentence-transformers
langchain
pdfminer.six
python-docx
boto3
pytest
requests
openai
transformers
torch
scikit-learn

langchain-community
langchain-core
pypdf
pymupdf
